pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "penguin-classifier"
version = "0.0.1"
description = "A platform-independent machine learning application to classify Palmer penguins, designed for non-technical field usage."
authors = [
    { name = "Erik Reitsam" },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10, <3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

dependencies = [
    "python-dotenv",
    "loguru",
    "numpy",
    "pandas",
    "scikit-learn",
    "dash",
    "dash-bootstrap-components",
    "seaborn >=0.10.0,<0.14.0",
    "matplotlib >=3.8.0,<3.10.0",
    "gunicorn >=25.0.3,<26.0.0",
    "joblib >=1.5.3,<2.0.0",
    "plotly >=6.5.2,<7.0.0",
]

[project.scripts]
penguin-app = "penguin_classifier.app:main"

[tool.poetry]
packages = [
    { include = "penguin_classifier", from = "src" }
]

[tool.poetry.group.dev.dependencies]
pytest = "*"
ruff = "*"
typer = "*"
palmerpenguins = "^0.1.6"

[tool.poetry.group.docs.dependencies]
mkdocs = "*"

[tool.poetry.group.eda.dependencies]
ipython = "*"
jupyterlab = "*"
notebook = "*"
tqdm = "*"
ipykernel = "^7.1.0"
ydata-profiling = "^4.18.0"

[tool.ruff]
line-length = 79
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["penguin_classifier"]
force-sort-within-sections = true